Feature main screen

+++ app/src/main/java/com/danilovfa/cryptocurrencies/data/remote/dto/CryptocurrenciesItemDto.kt
@@ -13,7 +13,7 @@ data class CryptocurrenciesItemDto(
     @SerializedName("image")
     val image: String,
     @SerializedName("market_cap")
-    val marketCap: Long,
+    val marketCap: Double,
     @SerializedName("name")
     val name: String,
     @SerializedName("price_change_24h")
